pyproject.toml
[project]
name = "qbuck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.7.14",
    "jupyterlab>=4.4.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pybind11>=3.0.1",
    "pylatexenc>=2.10",
    "python-dotenv>=1.2.1",
    "qiskit==2.1.1",
    "qiskit-aer>=0.17.1",
    "qiskit-ibm-provider>=0.11.0",
    "qiskit-ibm-runtime>=0.44.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "clang-format>=21.1.7",
    "cmake!=3.17.0,!=3.17.1",
    "conan<2.0.0",
    "scikit-build>=0.11.0",
]
